move enums to separate file
